refactor: Enhance MCP module with async configuration support and new core module (#50)
Changes:

- Refactored module structure for better organization
- Fixed forRootAsync imports and type definitions
- Corrected forRootAsync usage example
- Updated documentation for forRootAsync

Closes #48
Closes #49
