+++ examples/README.md
@@ -19,7 +19,7 @@ This directory contains ready-to-use example MCP servers demonstrating how to us
 - **guards/**  
   Shows how to implement and use guards in MCP servers to control access to capabilities and resources. Demonstrates different guard patterns and how to apply them to protect specific MCP features.
 
-- **async-import/**  
+- **for-root-async/**  
   Example of asynchronous module configuration of the MCP module using `forRootAsync` and environment variables.
 
 ## How to Run Examples
