+++ examples/for-root-async/main.ts
