+++ src/mcp.constants.ts
@@ -1,0 +1,3 @@
+export const MCP_MODULE_OPTIONS = 'MCP_MODULE_OPTIONS';
+export const MCP_LOGGING_OPTIONS = 'MCP_LOGGING_OPTIONS';
+export const MCP_TRANSPORT_OPTIONS = 'MCP_TRANSPORT_OPTIONS';
